+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Development Version (unreleased):
 
 - Fix unintentional changes in test suite (:pull:`277`).
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_
+- Create ENUM with low level API to keep order-by-value, add h5dump based tests (:pull:`285`).
+  By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_ and `David Huard <https://github.com/huard>`_
 - Add the `format` argument to `h5netcdf.File` and partial support for the `NETCDF4_CLASSIC` format (:issue:`280`, :pull:``).
   By `David Huard <https://github.com/huard>`_